requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
liac-arff==2.5.0
scipy==1.11.3
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9